Function declarator with no prototype problem solved (void is not specified); explicit cast to uint8_t in some assignment

+++ src/screen.c
@@ -3,7 +3,7 @@
 
 void pacman_effect(uint8_t* x, uint8_t* y) {
     if ((*x) < LO_BOUND_X) {
-        (*x) = HI_BOUND_X-HALF_SPRITE;
+        (*x) = (uint8_t)(HI_BOUND_X-HALF_SPRITE);
     } 
     else if ((*x) > HI_BOUND_X)
     {
@@ -11,7 +11,7 @@ void pacman_effect(uint8_t* x, uint8_t* y) {
     }
     
     if ((*y) < LO_BOUND_Y) {
-        (*y) = HI_BOUND_Y-HALF_SPRITE;
+        (*y) = (uint8_t)(HI_BOUND_Y-HALF_SPRITE);
     } 
     else if ((*y) > HI_BOUND_Y)
     {
